Restaurante/requirements.txt
Flask
flask-marshmallow
marshmallow
Flask-SQLAlchemy
SQLAlchemy
Flask-Migrate
psycopg2-binary
python-dotenv
sqlalchemy-easy-softdelete
marshmallow-sqlalchemy